+++ client/src/components/NavBar/index.js
@@ -1,113 +1,113 @@
-// import React from "react";
-// import { makeStyles } from "@material-ui/core/styles";
-// import AppBar from "@material-ui/core/AppBar";
-// import Toolbar from "@material-ui/core/Toolbar";
-// import Typography from "@material-ui/core/Typography";
-// import IconButton from "@material-ui/core/IconButton";
-// // import MenuIcon from "@material-ui/icons/Menu";
-// import AccountCircle from "@material-ui/icons/AccountCircle";
-// // import Switch from "@material-ui/core/Switch";
-// // import FormControlLabel from "@material-ui/core/FormControlLabel";
-// // import FormGroup from "@material-ui/core/FormGroup";
-// import MenuItem from "@material-ui/core/MenuItem";
-// import Menu from "@material-ui/core/Menu";
+import React from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import AppBar from "@material-ui/core/AppBar";
+import Toolbar from "@material-ui/core/Toolbar";
+import Typography from "@material-ui/core/Typography";
+import IconButton from "@material-ui/core/IconButton";
+// import MenuIcon from "@material-ui/icons/Menu";
+import AccountCircle from "@material-ui/icons/AccountCircle";
+// import Switch from "@material-ui/core/Switch";
+// import FormControlLabel from "@material-ui/core/FormControlLabel";
+// import FormGroup from "@material-ui/core/FormGroup";
+import MenuItem from "@material-ui/core/MenuItem";
+import Menu from "@material-ui/core/Menu";
 
-// //  import { Route } from "react-router";
+//  import { Route } from "react-router";
 
-// const useStyles = makeStyles((theme) => ({
-//   root: {
-//     flexGrow: 1,
-//   },
-//   // menuButton: {
-//   //   marginRight: theme.spacing(0),
-//   // },
-//   title: {
-//     flexGrow: 1,
-//   },
-// }));
+const useStyles = makeStyles((theme) => ({
+  root: {
+    flexGrow: 1,
+  },
+  // menuButton: {
+  //   marginRight: theme.spacing(0),
+  // },
+  title: {
+    flexGrow: 1,
+  },
+}));
 
-// const NavStlye = {
-//   background: "darkgreen",
-//   color: "black",
-//   fontvariant: "all-petite-caps",
-// };
+const NavStlye = {
+  background: "darkgreen",
+  color: "black",
+  fontvariant: "all-petite-caps",
+};
 
-// export default function NavBar() {
-//   const classes = useStyles();
-//   const [auth, setAuth] = React.useState(true);
-//   const [anchorEl, setAnchorEl] = React.useState(null);
-//   const open = Boolean(anchorEl);
+export default function NavBar() {
+  const classes = useStyles();
+  const [auth, setAuth] = React.useState(true);
+  const [anchorEl, setAnchorEl] = React.useState(null);
+  const open = Boolean(anchorEl);
 
-//   const handleChange = (event) => {
-//     setAuth(event.target.checked);
-//   };
+  const handleChange = (event) => {
+    setAuth(event.target.checked);
+  };
 
-//   const handleMenu = (event) => {
-//     setAnchorEl(event.currentTarget);
-//   };
+  const handleMenu = (event) => {
+    setAnchorEl(event.currentTarget);
+  };
 
-//   const handleClose = () => {
-//     setAnchorEl(null);
-//   };
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
 
-//   return (
-//     <div className={classes.root}>
-//       <AppBar position="static">
-//         <Toolbar style={NavStlye}>
-//           {/* <IconButton
-//             edge="start"
-//             className={classes.menuButton}
-//             color="inherit"
-//             aria-label="menu"
-//           >
-//             <MenuIcon />
-//           </IconButton> */}
-//           <Typography variant="h6" className={classes.title}>
-//             Gamedex
-//           </Typography>
-//           {/* <Typography onClick variant="h6" className={classes.title}>
-//             Steam
-//           </Typography>
-//           <Typography variant="h6" className={classes.title}>
-//             Epic
-//           </Typography> */}
-//           {auth && (
-//             <div>
-//               <IconButton
-//                 aria-label="account of current user"
-//                 aria-controls="menu-appbar"
-//                 aria-haspopup="true"
-//                 onClick={handleMenu}
-//                 color="inherit"
-//               >
-//                 <AccountCircle />
-//               </IconButton>
-//               <Menu
-//                 id="menu-appbar"
-//                 anchorEl={anchorEl}
-//                 anchorOrigin={{
-//                   vertical: "top",
-//                   horizontal: "right",
-//                 }}
-//                 keepMounted
-//                 transformOrigin={{
-//                   vertical: "top",
-//                   horizontal: "right",
-//                 }}
-//                 open={open}
-//                 onClose={handleClose}
-//               >
-//                 <MenuItem onClick={handleClose}>Profile</MenuItem>
-//                 <MenuItem onClick={handleClose}>My Wishlist</MenuItem>
+  return (
+    <div className={classes.root}>
+      <AppBar position="static">
+        <Toolbar style={NavStlye}>
+          {/* <IconButton
+            edge="start"
+            className={classes.menuButton}
+            color="inherit"
+            aria-label="menu"
+          >
+            <MenuIcon />
+          </IconButton> */}
+          <Typography variant="h6" className={classes.title}>
+            Gamedex
+          </Typography>
+          {/* <Typography onClick variant="h6" className={classes.title}>
+            Steam
+          </Typography>
+          <Typography variant="h6" className={classes.title}>
+            Epic
+          </Typography> */}
+          {auth && (
+            <div>
+              <IconButton
+                aria-label="account of current user"
+                aria-controls="menu-appbar"
+                aria-haspopup="true"
+                onClick={handleMenu}
+                color="inherit"
+              >
+                <AccountCircle />
+              </IconButton>
+              <Menu
+                id="menu-appbar"
+                anchorEl={anchorEl}
+                anchorOrigin={{
+                  vertical: "top",
+                  horizontal: "right",
+                }}
+                keepMounted
+                transformOrigin={{
+                  vertical: "top",
+                  horizontal: "right",
+                }}
+                open={open}
+                onClose={handleClose}
+              >
+                <MenuItem onClick={handleClose}>Profile</MenuItem>
+                <MenuItem onClick={handleClose}>My Wishlist</MenuItem>
 
-//                 {/* <Route exact path="/Wishlist" component={Wishlist}>
-//                   <MenuItem onClick={handleClose}>My Wishlist</MenuItem>
-//                 </Route> */}
-//               </Menu>
-//             </div>
-//           )}
-//         </Toolbar>
-//       </AppBar>
-//     </div>
-//   );
-// }
+                {/* <Route exact path="/Wishlist" component={Wishlist}>
+                  <MenuItem onClick={handleClose}>My Wishlist</MenuItem>
+                </Route> */}
+              </Menu>
+            </div>
+          )}
+        </Toolbar>
+      </AppBar>
+    </div>
+  );
+}
