+++ client/src/pages/wishlist.js
@@ -1,15 +1,14 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Wishlist from "../components/Wishlist/index";
 import Navbar from "../components/NavBar/index.js";
 
-
 function WishlistPage() {
-    return(
-        <div>
-        <Navbar />
-        <Wishlist />
-        </div>
-    )
-} 
+  return (
+    <div>
+      <Navbar />
+      <Wishlist />
+    </div>
+  );
+}
 
 export default WishlistPage;
